Added SceneParticle library.
Contains node to hold particles, prepare list to create and prepare
particle emitters, and draw list to draw dsParticleDraw instances. Still
need to hook up transforms and custom resource wrapper for dsParticleDraw
and particle emitter factories. The particle emitter factory function is
currently a placeholder and will change once the other parts are hooked up.

Added update function to dsSceneItemList to allow a once per scene update.
dsScene_update() now takes a time and updates the item lists. Application
update function now takes a float time rather than double, since it's
relative to the last frame and nearly all logic will be in floats
regardless.

Commit function is now optional for dsSceneItemList unless in a render
pass. This allows for scene item lists that only perform an update, though
an item list in a render pass only makes sense if it commits its contents.
(i.e. draw in this context)

+++ modules/Application/Application/include/DeepSea/Application/Types.h
@@ -588,7 +588,7 @@ typedef struct dsEvent
  * @param lastFrameTime The time it took to execute the last frame in seconds.
  * @param userData The user data registered with the function.
  */
-typedef void (*dsUpdateApplicationFunction)(dsApplication* application, double lastFrameTime,
+typedef void (*dsUpdateApplicationFunction)(dsApplication* application, float lastFrameTime,
 	void* userData);
 
 /**
